# apps/interlabs-crm-demo/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Interlab Portal API
  version: 1.0.0
  description: |
    Internal Enterprise Dashboard — CRM + ERP + Realtime Operations Hub for
    PT. Interlab Sentra Solutions Indonesia.

    Covers: Auth, Users & Roles, PO lifecycle (11 stages across 5 divisions),
    Sales, Admin & Log, Finance, Technical, HRGA/Legal, Tax & Insurance,
    Notifications, Chat, Email Templates, Files.

    ## Response envelope
    Every response uses a consistent envelope:
      success: `{ "success": true,  "data": ..., "meta"?: {...} }`
      error:   `{ "success": false, "error": "message", "code"?: "..." }`

    ## RBAC
    Every protected endpoint declares `x-rbac` with the feature_key,
    capability, allowed role_keys, and row-level scope. The RBAC middleware
    enforces `(feature, capability)` against `role_permissions` on every
    request; `scope` is applied to queries via the sameRoleScope /
    row-level-security layer.

    ## PO lifecycle side-effects
    Endpoints marked `x-po-side-effects` MUST, in a single transaction:
      1. Append to `purchase_order_status_history`
      2. Append to `purchase_order_tracking_events`
      3. Update `purchase_orders.current_status`
      4. Emit the declared notification templates (if enabled)
  contact:
    name: Interlab Portal Engineering
    email: zakyindrasatriaputra@gmail.com
servers:
  - url: https://api.interlab-portal.com
    description: Production
  - url: http://localhost:3000
...
security:
  - bearerAuth: []
paths:
...
  /api/po-tracking/search:
    get:
      tags: [PO Tracking]
      summary: Search PO by po_number
      x-rbac:
        feature: po_tracking
        capability: view_own
        allowed_roles: [superadmin, ceo, sales, admin_log, finance, technical, hrga, tax_insurance]
        scope: role-scoped
      parameters:
        - { in: query, name: po_number, required: true, schema: { type: string } }
      responses:
        '200':
          description: PO with latest 3 history entries
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data: { $ref: '#/components/schemas/POSearchResult' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/po-tracking/{id}/history:
    parameters:
      - $ref: '#/components/parameters/UUIDPathId'
    get:
      tags: [PO Tracking]
      summary: Full PO stage-transition history
      x-rbac:
        feature: po_tracking
        capability: view_own
        allowed_roles: [superadmin, ceo, sales, admin_log, finance, technical, hrga, tax_insurance]
        scope: role-scoped
      responses:
        '200':
          description: History entries (oldest -> newest)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/POStatusHistoryEntry' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/notifications:
    get:
      tags: [Notifications]
      summary: Notification bell feed (latest unread by default)
      x-rbac: { feature: notifications, capability: view_own, allowed_roles: [superadmin, ceo, sales, admin_log, finance, technical, hrga, tax_insurance], scope: own }
      parameters:
        - { in: query, name: limit, schema: { type: integer, default: 5 } }
        - { in: query, name: unread, schema: { type: boolean, default: true } }
      responses:
        '200':
          description: Notifications
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Notification' }
                      meta:
                        type: object
                        properties:
                          unread_count: { type: integer }
  /api/notifications/all:
    get:
      tags: [Notifications]
      summary: Full paginated notifications page
      x-rbac: { feature: notifications, capability: view_own, allowed_roles: [superadmin, ceo, sales, admin_log, finance, technical, hrga, tax_insurance], scope: own }
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - { in: query, name: unread, schema: { type: boolean } }
        - { in: query, name: related_module, schema: { type: string } }
      responses:
        '200':
          description: Paginated notifications
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Notification' }
  /api/notifications/{id}/read:
    parameters: [ { $ref: '#/components/parameters/UUIDPathId' } ]
    put:
      tags: [Notifications]
      summary: Mark notification as read
      x-rbac: { feature: notifications, capability: edit, allowed_roles: [superadmin, ceo, sales, admin_log, finance, technical, hrga, tax_insurance], scope: own }
      responses:
        '200':
          description: Marked as read
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessEnvelope' }
...
  /api/files/upload:
    post:
      tags: [Files]
      summary: Upload a file to MinIO
      description: |
        Accepts multipart/form-data. Supported extensions:
        pdf, doc, docx, xls, xlsx, csv, jpg, jpeg, png, webp, zip.
        Server assigns file_id and stores metadata in file_attachments.
      x-rbac: { feature: user_profile, capability: write, allowed_roles: [superadmin, ceo, sales, admin_log, finance, technical, hrga, tax_insurance], scope: own }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, related_module]
              properties:
                file:
                  type: string
                  format: binary
                related_module:
                  type: string
                  enum: [sales, admin-log, finance, technical, hrga, tax, po-tracking, avatars]
                related_entity_id:
                  type: string
                  format: uuid
      responses:
        '201':
          description: File stored + metadata row created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data: { $ref: '#/components/schemas/FileAttachment' }
        '413': { description: File too large }
  /api/files/{id}/presigned-url:
    parameters: [ { $ref: '#/components/parameters/UUIDPathId' } ]
    get:
      tags: [Files]
      summary: Get 15-minute presigned download URL
      x-rbac: { feature: user_profile, capability: view_own, allowed_roles: [superadmin, ceo, sales, admin_log, finance, technical, hrga, tax_insurance], scope: own }
      responses:
        '200':
          description: Presigned URL
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data: { $ref: '#/components/schemas/FilePresignedUrl' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/files/{id}:
    parameters: [ { $ref: '#/components/parameters/UUIDPathId' } ]
    delete:
      tags: [Files]
      summary: Soft-delete file (removes MinIO object + marks deleted_at)
      x-rbac: { feature: user_profile, capability: delete, allowed_roles: [superadmin, ceo, sales, admin_log, finance, technical, hrga, tax_insurance], scope: own }
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token (1h expiry). Obtained from `POST /api/auth/login`.
        Refresh with `POST /api/auth/refresh` (refresh token 7d, or 30d with remember_me).
  parameters:
    PageParam:
      in: query
      name: page
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
    LimitParam:
      in: query
      name: limit
      required: false
      schema: { type: integer, enum: [25, 50, 75, 100], default: 25 }
...
    UUIDPathId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
  responses:
...
    NotFound:
      description: Resource not found (or soft-deleted)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, enum: [true] }
        data: {}
        meta: { $ref: '#/components/schemas/PaginationMeta' }
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, enum: [false] }
        error: { type: string }
        code: { type: string }
        field_errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
    PaginationMeta:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        total_pages: { type: integer }

    # ------------------------------------------------------------------
    # Enums
    # ------------------------------------------------------------------
...
    POStatus:
      type: string
      enum: [Registered, Processed, Production, Shipped, Customs, Arrived, Inspected, Delivery, Installation, BAST, Invoice]
    POStatusCode:
      type: string
      enum: [REGISTERED, PROCESSED, PRODUCTION, SHIPPED, CUSTOMS, ARRIVED, INSPECTED, DELIVERY, INSTALLATION, BAST, INVOICE]
...
    PurchaseOrder:
      type: object
      properties:
        id: { type: string, format: uuid }
        po_number: { type: string }
        current_status: { $ref: '#/components/schemas/POStatus' }
        created_by_user_id: { type: string, format: uuid, nullable: true }
        created_by_role: { type: string, nullable: true }
        updated_by_user_id: { type: string, format: uuid, nullable: true }
        updated_by_role: { type: string, nullable: true }
        customer_id: { type: string, format: uuid, nullable: true }
        due_at: { type: string, format: date-time, nullable: true }
        overdue_at: { type: string, format: date-time, nullable: true }
        overdue_reason: { type: string, nullable: true }
        overdue_attachment_id: { type: string, format: uuid, nullable: true }
        escalation_sent_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    POStatusHistoryEntry:
      type: object
      properties:
        id: { type: string, format: uuid }
        po_id: { type: string, format: uuid }
        po_number: { type: string }
        status_code: { $ref: '#/components/schemas/POStatusCode' }
        status_label: { type: string }
        updated_by_user_id: { type: string, format: uuid, nullable: true }
        updated_by_role: { type: string, nullable: true }
        note: { type: string, nullable: true }
        reason_if_delayed: { type: string, nullable: true }
        attachment_url: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
...
    POSearchResult:
      type: object
      description: Role-scoped PO search result — includes latest 3 history entries
      properties:
        po: { $ref: '#/components/schemas/PurchaseOrder' }
        latest_history:
          type: array
          maxItems: 3
          items: { $ref: '#/components/schemas/POStatusHistoryEntry' }
...
    FileAttachment:
      type: object
      properties:
        id: { type: string, format: uuid }
        original_filename: { type: string }
        mime_type: { type: string }
        extension: { type: string, nullable: true }
        size_bytes: { type: integer, nullable: true }
        uploaded_by: { type: string, format: uuid, nullable: true }
        uploaded_at: { type: string, format: date-time }
        related_module: { type: string }
        related_entity_id: { type: string, format: uuid, nullable: true }
        storage_bucket: { type: string }
        storage_path: { type: string }
    FilePresignedUrl:
      type: object
      properties:
        url: { type: string }
        expires_at: { type: string, format: date-time }

    # ------------------------------------------------------------------
    # Sales — forecasts, quotations, HPP, PO, PR
    # ------------------------------------------------------------------
...
    Notification:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        message: { type: string, nullable: true }
        recipient_user_id: { type: string, format: uuid, nullable: true }
        recipient_role: { type: string, nullable: true }
        sender_user_id: { type: string, format: uuid, nullable: true }
        related_module: { type: string, nullable: true }
        related_entity_type: { type: string, nullable: true }
        related_entity_id: { type: string, format: uuid, nullable: true }
        is_read: { type: boolean }
        created_at: { type: string, format: date-time }
...